Consent module.
This implements a module for requesting attribute release consent from the
user. It is based on the consent support already in simpleSAMLphp. There are
some changes brom the builtin consent support. The database schema is
different, the table now includes a UNIQUE constraint on user and destination,
and the destination is no longer encoded as a hash.

There is also support for saving the consent in cookies.

See modules/consent/Auth/Process/Consent.php for information about
configuration.