Add support for sending responses with the HTTP-Artifact binding.
This patch implements support for sending responses to authentication
requests via the HTTP-Artifact binding. To enable, add
'saml20.sendartifact' => TRUE in saml20-idp-hosted metadata. The IdP
should then send HTTP-Artifact responses to SPs that request it.

Note that this requires a working memcache server.

Thanks to Danny Bollaert for implementing support for this.