Updated the casserver module as per Dubravko Voncinas comments in http://rnd.feide.no/content/simplesamlphp-cas-server-support.
cas.php is a frontend for login.php, validate.php and
serviceValidate.php. It allows them to be called as cas.php/login,
cas.php/validate and cas.php/serviceValidate and is meant for clients
like phpCAS which expects one configured prefix to which it appends
login, validate and serviceValidate.

The apache configurationchanges Dubravko Voncina mention is
thus not necessary any longer.