shibmd:scope regexp attribute should default to false and be always explicit
The regexp attribute of a Scope used to default to null when not set
explicitly.

According to the spec, when omitted this should be interpreted as it being
false. Therefore, I think the null default is not correct to have.

Also, we should always output an explicit regexp attribute,
as recommended by the spec for document signing reasons:
https://wiki.shibboleth.net/confluence/display/SC/ShibMetaExt+V1.0