Be tolerant of string EPTIs instead of breaking mere parsing of those assertions.
Before, parsing would just break with an exception if the IdP sent such
a string value, even if the SP would do nothing with this attribute at
all. The SP cannot always control the IdP version, which can be an
implementation that sends the wrong string value, since they are
somewhat common, not in the least SSP itself. Instead, now we log a
warning but continue processing.