Trim extraneous whitespace on URLs before validation
It is not uncommon for logo URLs in metadata feeds to include extraneous
whitespace, and most user interfaces and other functions making use of
the MDUI data handle this gracefully. However, FILTER_VALIDATE_URL
regards this as an error and rejects the URL.

Per the robustness principle (Postel's law), it probably makes sense
to accept these, rather than to reject an entire feed because of some
extraneous whitespace that probably doesn't matter anyway.

RFC 3986 does not allow any of the characters removed by PHP's trim()
function to be used without percent encoding. So wrapping the URL
in a trim() should be a no-op for any URL that is well formed.