Improve exception messages
I attempted to improve missing or unclear exception messages. My
assumption that the SAML2 library contained hard to grasp exceptions was
not valid at all. Most messages are very clear. It might depend on the
way you use the library, how much sense the message makes in the context
you are using it in, but that should be fixed by the implementor and not
by us.