Add static methods to AbstractXMLElement to get attributes from a given element
For now, this covers strings, integers and booleans. Passing a string value as the last argument gets it used as a default if the attribute is not defined. Passing null as the last argument makes it nullable, so that its result can be plugged into a setter method. Omit the $default argument and the attribute becomes mandatory.