Simplify extensions
We had quite a mess. This commit simplifies the extension system by leaving the ExtendableElement interface to the most basic functionality (declaring a property and accessing it), and moving actual functionality back to SAML2\XML\md\Extensions. This class should have always represented an md:Extensions element, but for some reason it didn't, making it more complicated, and leaving us with its own API. Here we reimplement it to extend AbstractMdElement, so that we can use fromXML() and toXML() to manage extensions, as we would do with any other element.

Note that ExtensionsTest hasn't still been adapted to the new implementation.