Add mailto:-prefix in EmailAddress property, if missing.
This is mandatory in the spec. It seems the role of the library to
add it, so users can continue to supply and receive bare email
addresses. The code is defensive in that it will also accept
incoming EmailAddress values that already sport the mailto:-prefix.

Also improve code coverage of ContactPerson