Remove checks for validUntil & cacheDuration
Unfortunately, we cannot perform this check in the library. The standard makes either one of them mandatory when the element is the root of a metadata instance. However, there's no way for us to know if that's the case, except for EntitiesDescriptor. EntityDescriptor can be inside an EntitiesDescriptor, and therefore we don't know if it's the root element or not. For the shake of consistency, I think we shouldn't perform the check in EntitiesDescriptor either, and leave that responsibility (maybe we should document this?) to the user of the library.