Rewrite api saml (#213)
* Migrate Issuer/NameID/NameIDType/BaseIDType

* Loosen typehints

* Some improvements

* Clear saml/Issuer of SPID specifics

* Fix merge issue

* Private setters

* Add protecting assertions

* Get rid of static call at the cost of code duplication

* Rid of undefined BaseIDType & smarten tests

* Add serialization-tests

* Improve tests

* Relax the strictness of Issuer

Instead of raising an exception if optional attributes are present for an Issuer whose format is "entity", just ignore those attributes and set the corresponding properties to null. This avoids an entity consuming a message conforming SPID from breaking.

* Simplify the Issuer implementation for SPID

Format is fixed, and the only variables that we are supposed to support are the value itself and NameQualifier. Use only those two and ignore the rest.

* Undo spid specifics; spid-support has been moved to saml2-module-spid

* Sanitize tests

* Fix NameIDType typehints

* Revert Issuer/NameID <> NameIDType

* Fix test

* Fix syntax

Co-authored-by: Jaime Pérez Crespo <jaime.perez@uninett.no>