Bugfix DecodeBase4Transformer and add tests.
The current code was not working because the decodeValue() function
makes an array from the input string so this has to be merged back
properly.

Also run base64_decode() in strict mode now so we get an exception
when the attribute value is not actual base64.

And add full test coverage.