First attempt (work in progress) to refactor the Assertion class
Most of the functionality is now in other classes, which are included here as objects. The assertion class is now responsible exclusively for its own children, but not for anything specific to them (e.g. attributes, attribute formats or attribute values). We still need to go through some parts of the code and evaluate whether the implementation here needs to be moved or adapted to the new classes.