Remove base64 encoded attributes transformer.
This would decode all received attribute values with base64. As far
as I'm aware this is from the early days of federation. In any case
SimpleSAMLphp (although it has/had a similar feature) does not use
this code. This major release seems like a fine moment to drop this
unless someone steps forward with an actual use case.